requirements.txt
discord.py==0.16.12
nltk==3.5
